Unstyled version ready for Heroku deployment
Everything should be set up and working. By default hardcoded values
will be used in development, instead of the real API.

The environment variables need to be set in Heroku.

heroku config:set THECITY_SECRET_KEY=""
heroku config:set THECITY_USER_KEY=""

The real values are in thecity.secret.exs but are not committed. They
are present so the real api can be used during development, but by
default hardcoded values are returned, instead of the actual API.