Rework logic when determining if a new page of checkins is needed
Instead of using a count the logic used to determine if a new page of
checkins is needed by first keeping the checkins as a list and checks if
the last one from the page of results just retrieved already exists in
the kids map. If not then a new page of checkins in requested.