Fix: Cache may never be updated on 0 count
If the last page of kids results happen to be the last new kid then a
new request to retrieve kids will be sent, but the response will show 0
new kids. This means the list of kids will not be cached. Now we always
update the cache, whether or not their are new kids in the last request.